app/requirements.txt
fastapi
uvicorn
opencv-python-headless
torch
torchvision
numpy
Pillow
requests
python-multipart